:bug: Fix variant auto-selection on load.
I switched to a custom option selector for the option dropdowns now,
this should make it easier to modify in the future, rather than relying
on Shopify's `option_selection.js` snippet.

There was a bug where we were triggering a `change`, which reset the
dropdown to the first item. I don't think this code snippet is needed
anymore.